Completed first attempt at indxex.js.
Installed required packages and node inclusions.  Added the packages needed @ the top per TODO.  Created an array of questions that will be used to complete the README.md. Created a initial functions to write the README and to initialize the application.

Currently a bug that only records the title to the text document.  Could be due to setup of the generateMarkdown.js call without completing that section of the assignment yet.  Will explore further.